refactor(layout): updated plugin file layout (#27)
