+++ sources/activity_log_api/activity_log_api_source_config.go
@@ -1,4 +1,4 @@
-package sources
+package activity_log_api
 
 type ActivityLogAPISourceConfig struct {
 }
