+++ sources/blob_storage/azure_blob_storage_source_config.go
@@ -1,4 +1,4 @@
-package sources
+package blob_storage
 
 import (
 	"fmt"
